requirements.txt
# FastAPI and ASGI server
fastapi>=0.100.0
uvicorn[standard]>=0.20.0

# Database
sqlalchemy>=2.0.0
alembic>=1.10.0
psycopg2-binary>=2.9.0
aiosqlite>=0.19.0  # Keep for testing

# Redis and background jobs
redis>=5.0.0
celery>=5.3.0

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Validation and serialization
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0

# HTTP client
httpx==0.25.2

# Email service
resend==0.6.0

# Utilities
python-dotenv==1.0.0
structlog==23.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
PyYAML>=6.0
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Template engine
Jinja2>=3.1.0

# System monitoring
psutil>=5.9.0

# AWS SDK for document storage
boto3>=1.26.0

# PDF generation (optional)
# xhtml2pdf>=0.2.5
